requirements.txt
fastapi
uvicorn
python-multipart
tensorflow==2.10.0
numpy
Pillow
opencv-python